pyproject.toml
[project]
name = "ndl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cassandra-driver>=3.29.2",
    "dateutils>=0.6.12",
    "dlib>=19.24.6",
    "face-recognition>=1.3.0",
    "flask-socketio>=5.3.7",
    "flask>=3.0.3",
    "numpy>=2.1.1",
    "opencv-python>=4.10.0.84",
    "paho-mqtt>=2.1.0",
    "pillow>=10.4.0",
    "pycryptodome>=3.20.0",
]
